requirements.txt
Pillow
rembg
numpy
flask
gunicorn